requirements.txt
# Xinmeng M87 Linux RGB – Python dependencies
# Install with: pip install -r requirements.txt

# ── Core (required) ─────────────────────────────────────────────────────────
hidapi>=0.14.0          # Linux HID device access (Phase 1, 4, 5)
pyusb>=1.2.1            # USB device detection fallback (Phase 1)

# ── Packet analysis (Phase 3) ────────────────────────────────────────────────
scapy>=2.5.0            # Pcap file parsing
# OR: pyshark>=0.6      # Alternative (needs tshark installed separately)

# ── Audio reactive RGB (Phase 6) ────────────────────────────────────────────
sounddevice>=0.4.6      # Audio input capture
numpy>=1.24.0           # FFT / signal processing

# ── GUI (optional) ──────────────────────────────────────────────────────────
PyQt5>=5.15.0           # Graphical control panel
